Fixed issue with using .cell(....) and a new page being created.
When a new page was created because of the document length causing a call to .addPage, the default value for 'h' was left as 'undefined'. That resulted in the new 'y' value being set to Nan. The resulting PDF had errors when it was opened. This is a fix for that issue. Now when a new page is created, the table will start at the initial y position as it did on the first page. This allows for any header information on the page to be preserved.