Update cell.js (#1)
When invoking the jsPDF.table() method a "TypeError: config.css is undefined" is thrown when there is no font-size property on the css property of the config object that is passed. In this case execution halts.
Added check for config.css and config.css['font-size']. Used a negative typeof check for "undefined" instead of a positive one for "number" to respect loose typing.