Add Graphics State Stack Operator
Add Graphics State Stack Operator For Clipping

There is no way to enlarge the current clipping path or to set a new
clipping path
without reference to the current one. However, since the clipping path
is part of
the graphics state, its effect can be localized to specific graphics
objects by enclosing
the modification of the clipping path and the painting of those objects
between a pair of q and Q operators (see Section 4.3.1, “Graphics State
Stack”).
Execution of the Q operator causes the clipping path to revert to the
value that
was saved by the q operator, before the clipping path was modified.