Fix loading image from path and add test
This node package implementing atob does not throw if the string is not base64,
thus making a call like `.addImage('path.jpg', ...`) render blank.

I wasn't sure on what the proper fix here, so I conditioned the call to the
conversion with a validation on the string. Added a test, though I wasn't
really able to run the tests locally. I'm not even sure why they were failing.