Auto paging now uses last element offset for specific page
Auto paging now remembers last element offset for all pages and uses correct one depending on page to prevent weird text offset. This happened because html2canvas doesn't send elements in order so it is possible that element from page 4 will arrive before element from page 1. Specifically in added test `<em>basic></em>` and `<strong>tiptap</strong>` will be processed as last elements for this pdf.
Note that this also means that it should be probably possible to create element which will not correctly increase offset (example: if above tiptap would need to be moved to next page it will cause that offset for pages 2-4 will be wrong).
Note also that based on how pages are retrieved it should be also possible that these elements might be positioned on wrong page (example: if we would have 100 pages that would increase offset by more than a page getPagesByPath would return +1 page for tiptap even if it should be on first page)